refactor: useSteps custom hook 분리
